ugrep: updated to 3.11.2
ugrep v3.11.2

This update includes usability improvements. We improved option --hyperlink to optionally take parameters PREFIX and +. The PREFIX parameter replaces file:// in the link with a custom PREFIX://. The + included the line number in the link, and when ugrep option -k is used, also includes the column number in the link.