py-mistune: update to 2.0.5.
pkgsrc note: the CVE listed below was only for 2.x, so pkgsrc is
not affected.

2.0.5

Improve on parsing list. Make it possible to customize list regex. ref #331

2.0.4

Fix url plugin in <a> tag
Fix * formatting


2.0.3

Fix table plugin
Security fix for CVE-2022-34749

2.0.2

cape_url via #295

2.0.1

Fix XSS for image link syntax.

2.0.0

First release of Mistune v2.

2.0.0rc1

In this release, we have a Security Fix for harmful links.

2.0.0alpha1

This is the first release of v2. An alpha version for users to have a preview of the new mistune.