py-pylint-plugin-utils: updated to 0.8.2
0.8.2

Removing the LICENSE and README files from the project include section, as this
had unintended consequences of installing them into the site-packages directory
instead of along side the package code. This is related to
python-poetry/poetry-2015 . Since the LICENSE is available in the dist-info
directory along side the package, then this is sufficient.

Additionally, the python version classifiers will be automatically added by
poetry from the version range specification in the tool.poetry.dependencies
section, so no need to explicitly define them.